+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to the "nspin" package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [1.4.0] - 2025-03-26
+
+### Changed
+
+- Migrate build process to `vite` for improved performance and modern features.
+
+### Removed
+
+- Remove `tsup` as the build tool and replace it with `vite`.
+
 ## [1.3.0] - 2025-03-25
 
 ### Added
@@ -49,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the npm package.
 
-[Unreleased]: https://github.com/ManuelGil/nspin/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/ManuelGil/nspin/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/ManuelGil/nspin/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/ManuelGil/nspin/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/ManuelGil/nspin/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/ManuelGil/nspin/compare/v1.1.0...v1.1.1